requirements.txt
web3==7.0.0
eth_account==0.13.2
python-dotenv==1.0.1